feat: Refactor error handling and response types across action files
