+++ apps/web/actions/plans.action.ts
@@ -4,19 +4,9 @@ import { Plan } from "@repo/db";
 import { planService } from "@repo/services";
 import { revalidatePath } from "next/cache";
 import { auth } from "@/lib/next-auth.lib";
-
-interface ErrorResponse {
-  error: string;
-  code: string;
-  status: number;
-}
-
-type ActionResponse<T> =
-  | { data: T; error: null }
-  | { data: null; error: ErrorResponse };
+import { ActionResponse } from "@/lib/types";
 
 interface IAddUserToPlan {
-
   planId: string;
 }
 
@@ -24,7 +14,7 @@ export async function addUserToPlan(data: IAddUserToPlan): Promise<boolean> {
   const session = await auth();
   const userId = session?.user?.id;
   const planId = data.planId;
-    revalidatePath("/dashboard/profile");
+  revalidatePath("/dashboard/profile");
   try {
     await planService.addUserToPlan(planId, String(userId));
     return true;
